+++ sample/src/main/java/com/synerise/sdk/sample/ui/dev/apiAdapter/fragments/client/ClientRequestEmailChangeFragment.java
@@ -19,7 +19,7 @@ import com.synerise.sdk.sample.ui.dev.BaseDevFragment;
 public class ClientRequestEmailChangeFragment extends BaseDevFragment {
 
     private IApiCall apiCall;
-    private TextInputLayout inputEmail, inputPassword, inputUuid;
+    private TextInputLayout inputEmail, inputPassword;
 
     public static ClientRequestEmailChangeFragment newInstance() {
         return new ClientRequestEmailChangeFragment();
@@ -37,9 +37,7 @@ public class ClientRequestEmailChangeFragment extends BaseDevFragment {
 
         inputEmail = view.findViewById(R.id.input_email);
         inputPassword = view.findViewById(R.id.input_password);
-        inputUuid = view.findViewById(R.id.input_uuid);
         view.findViewById(R.id.email_change).setOnClickListener(v -> requestEmailChange());
-        view.findViewById(R.id.generate_uuid).setOnClickListener(this::onGenerateUuidClicked);
     }
 
     @Override
@@ -52,27 +50,20 @@ public class ClientRequestEmailChangeFragment extends BaseDevFragment {
     private void requestEmailChange() {
 
         inputPassword.setError(null);
-        inputUuid.setError(null);
 
         boolean isValid = true;
 
         String email = inputEmail.getEditText().getText().toString();
         String password = inputPassword.getEditText().getText().toString();
-        String uuid = inputUuid.getEditText().getText().toString();
 
         if (TextUtils.isEmpty(password)) {
             isValid = false;
             inputPassword.setError(getString(R.string.error_empty));
         }
 
-        if (TextUtils.isEmpty(uuid)) {
-            isValid = false;
-            inputUuid.setError(getString(R.string.error_empty));
-        }
-
         if (isValid) {
             if (apiCall != null) apiCall.cancel();
-            apiCall = Client.requestEmailChange(email, password, uuid);
+            apiCall = Client.requestEmailChange(email, password, null, null);
             apiCall.execute(this::onSuccess, new DataActionListener<ApiError>() {
                 @Override
                 public void onDataAction(ApiError apiError) {
@@ -80,10 +71,5 @@ public class ClientRequestEmailChangeFragment extends BaseDevFragment {
                 }
             });
         }
-    }
-
-    @SuppressWarnings("ConstantConditions")
-    private void onGenerateUuidClicked(View v) {
-        inputUuid.getEditText().setText(Client.getUuid());
     }
 }